add map on show pet
