+++ app/views/pets/show.html.erb
@@ -10,18 +10,11 @@
   <div class="row">
     <div class="col-sm-12 col-md-6">
       <table class="table table-bordered">
-        <tbody>
-          <thead>
-            <th>Name</th>
-            <th>Species</th>
-            <th>Name of my owner</th>
-          </thead>
-            <tr>
-              <td><%= @pet.name.capitalize %></td>
-              <td><%= @pet.species %></td>
-              <td><%= @pet.pet_owner.name.capitalize %></td>
-            </tr>
-        </tbody>
+        <ul>
+          <li>Name : <%= @pet.name.capitalize %></li>
+          <li>Species : <%= @pet.species %></li>
+          <li>Name of my owner : <%= @pet.pet_owner.name.capitalize %></li>
+        </ul>
       </table>
     </div>
   </div>
